gnu-linux: ignore shellcheck warning
Shellcheck is a static analyzer for shell script. It helps you avoid
making silly mistakes that can be hard to find. Check it out here:
https://u27125094.ct.sendgrid.net/ls/click?upn=Y73ls-2FpqpPk8ipT8ZhG2-2FiJYe-2FjtN70wBMZydtGEpCHAVZsNVxE9JkeOE8-2BavGqoA-2FEd-2FVmtXYKZv-2FPbcdI5Fg-3D-3D-oKI_XjzyZYWgSQeZavOB6BObZ9cPKBJyq2WB2rtE-2B8bIpruHZ-2F9H6Qv-2ByTU5Jovqej-2BlyrTyjX4UU33GvUwNu8g9DQ26mgX64AT9cigmJnGlsgMTB5ORZ1YFVVuEck5KcwzDcrA5aYM12ZxkCdCc-2FvmrjFpmb2AC68mZZqD-2BoJXEOj4v31Y7oDpQyJiTtwy3rnjk2oDdLiwx4WFW5vH21ja05iVfuIyhH58fcyKEG90vHf0-3D

This shellcheck warning can be ignored because assigning a variable
should always be safe to do, the only reason I could think it wouldn't
succeed is if the variable was read-only, and I've only seen that used
with docker containers.